requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6

# 深度学习
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0

# 向量检索
faiss-cpu>=1.7.4

# 图像处理
Pillow>=8.3.0
scipy>=1.7.0

# 数据处理
pandas>=1.3.0

# 其他工具
six>=1.16.0 